# fix: change logic of test production and consumption metabolites to speed up the process