--- a/src/memote/support/consistency.py
+++ b/src/memote/support/consistency.py
@@ -23,9 +23,11 @@ import logging
 from operator import attrgetter
 
 import numpy as np
+import multiprocessing
 from cobra import Reaction
 from cobra.exceptions import Infeasible
 from cobra.flux_analysis import flux_variability_analysis
+from cobra import Configuration
 from optlang.interface import OPTIMAL, INFEASIBLE
 from optlang.symbolics import Zero
 
@@ -490,37 +492,51 @@ def find_disconnected(model):
     return [met for met in model.metabolites if len(met.reactions) == 0]
 
 
-def solve_boundary(metabolite, rxn, val=-1):
+def _init_worker(model, irr):
+    """Initialize a global model object for multiprocessing.
+
+    Parameters
+    ----------
+    model : cobra.Model
+        The metabolic model under investigation.
+    irr: optlang.Variable || cobra.Reaction
+        the reaction to be added to the linear coefficients. It must be in the
+        variables of the model.
     """
-    Solves the model when some reaction `rxn` has been added to a `metabolite`'s
-    contraints.
+    global _model
+    global _irr
+    _model = model
+    _model.objective = irr
+    _irr = irr
+
+
+def _solve_metabolite_production(metabolite):
+    """
+    Solves the model when some reaction has been added to a `metabolite`'s
+    contraints. The reaction and the model are passed as globals.
 
     Parameters
     ----------
     metabolite: cobra.Metabolite
         the reaction will be added to this metabolite as a linear coefficient
-    rxn: optlang.Variable || cobra.Reaction
-        the reaction to be added to the linear coefficients. It must be in the
-        variables of the model.
-    val: float
-        the value of the linear coefficient of `rxn` in `metabolite`
 
     Returns
     -------
     solution: float
-        The value of the solution of the LP problem, *NaN* if infeasible.
-
+        the value of the solution of the LP problem, *NaN* if infeasible.
+    metabolite: cobra.Metabolite
+        metabolite passed as argument (to use map as a filter)
     """
-    metabolite.constraint.set_linear_coefficients({rxn: val})
-    # TODO: check if the "try except is required for older versions of cobra
-    solution = metabolite.model.slim_optimize()
-    # TODO: it seems like with context doesn't catch these changes, check
-    # restore constraint
-    metabolite.constraint.set_linear_coefficients({rxn: 0})
-    return solution
+    global _model
+    global _irr
+    constraint = _model.metabolites.get_by_id(metabolite.id).constraint
+    constraint.set_linear_coefficients({_irr: -1})
+    solution = _model.slim_optimize()
+    constraint.set_linear_coefficients({_irr: 0})
+    return solution, metabolite
 
 
-def find_metabolites_not_produced_with_open_bounds(model):
+def find_metabolites_not_produced_with_open_bounds(model, processes=None):
     """
     Return metabolites that cannot be produced with open exchange reactions.
 
@@ -534,6 +550,8 @@ def find_metabolites_not_produced_with_open_bounds(model):
     ----------
     model : cobra.Model
         The metabolic model under investigation.
+    processes: int
+        Number of processes to be used (Default to `cobra.Configuration()`).
 
     Returns
     -------
@@ -541,18 +559,34 @@ def find_metabolites_not_produced_with_open_bounds(model):
         Those metabolites that could not be produced.
 
     """
-    mets_not_produced = list()
+    if processes is None:
+        # For now, borrow the number of processes from cobra's configuration
+        processes = Configuration().processes
+    n_mets = len(model.metabolites)
+    processes = min(processes, n_mets)
+
     helpers.open_exchanges(model)
-    # custom Variable (demand reaction)
     irr = model.problem.Variable("irr", lb=0, ub=1000)
-    with model:
-        model.add_cons_vars(irr)
-        # helper.run_fba() only accepts reactions in the model
-        model.objective = irr
-        for met in model.metabolites:
-            solution = solve_boundary(met, irr)
-            if np.isnan(solution) or solution < model.tolerance:
-                mets_not_produced.append(met)
+
+    if processes > 1:
+        chunk_s = n_mets // processes
+        pool = multiprocessing.Pool(
+            processes,
+            initializer=_init_worker,
+            initargs=(model, irr),
+        )
+        # use map as filter
+        mets_not_produced = [met for solution, met in pool.imap_unordered(
+            _solve_metabolite_production, model.metabolites, chunksize=chunk_s
+        ) if np.isnan(solution) or solution < model.tolerance]
+        pool.close()
+        pool.join()
+    else:
+        _init_worker(model, irr)
+        # use map as filter
+        mets_not_produced = [met for solution, met in map(
+            _solve_metabolite_production, model.metabolites
+        ) if np.isnan(solution) or solution < model.tolerance]
     return mets_not_produced