In order to speed up the consistency tests of checking metabolite production and consumption with open bounds, the logic of the test has been changed. Instead of adding a boundary to the model each time, the test now works by sequentially changing the linear coefficients of the metabolites, which should be faster.

* Add a variable **just once** (bounds in 0,1000);
* Set the variable as the objective of the model;
* for each metabolite:  
        ▪ get the linear coefficients of the constraint of each metabolite;  
        ▪ add a coefficient to the variable in the constraint (-1 for production and 1 for consumption)
        ▪ solve;  
        ▪ restore constraint;  
        ▪ If not solution, add the metabolite to the returned output.

**Multiprocessing** was also added to decrease the running time.