@ChristianLieven 2019-11-14
I like your changes especially the use of mutliprocessing and manipulating the underlying model directly.

I'd be interested to see how much faster this is. Would you mind providing a before/after comparison?

And, of course, I'd feel more comfortable merging it if the test wouldn't fail.